refactor: Cell, Cells 파일 분리
