+++ src/main/kotlin/minesweeper/domain/cell/Cell.kt
@@ -1,8 +1,5 @@
 package minesweeper.domain.cell
 
-import minesweeper.domain.board.NearbyMineCounter
-import minesweeper.domain.cell.Cell.Empty
-
 sealed class Cell(val position: Position) {
 
     class Mine(position: Position) : Cell(position)
@@ -27,45 +24,4 @@ sealed class Cell(val position: Position) {
     fun isClosed() = this.state == CellStatus.CLOSE
 
     fun isOpen() = this.state == CellStatus.OPEN
-}
-
-class Cells(private val cells: List<Cell>) : List<Cell> by cells {
-
-    fun open(position: Position): Cell {
-        val cell = findCell(position).open()
-        if (cell is Empty) {
-            cell.openNearbyCells()
-        }
-        return cell
-    }
-
-    fun hasClosedEmptyCell() = cells.any { it is Empty && it.isClosed() }
-
-    private fun Empty.openNearbyCells() {
-        if (this.numberOfNearbyMines != 0) {
-            return
-        }
-        this.getNearbyCells().forEach { nearbyCell ->
-            nearbyCell
-                .takeIf { nearbyCell is Empty && nearbyCell.isClosed() }
-                ?.let { open(nearbyCell.position) }
-        }
-    }
-
-    fun Cell.getNearbyCells(): Cells {
-        val nearbyPositions = cells
-            .map { it.position }
-            .intersect(this.position.getNearbyPosition())
-        return Cells(nearbyPositions.map { findCell(it) })
-    }
-
-    private fun findCell(position: Position): Cell =
-        cells.find { it.position == position } ?: throw IllegalArgumentException("no cells found in that position.")
-
-    companion object {
-        fun of(width: Int, height: Int, numberOfMines: Int): Cells {
-            val cells = CellMaker.make(width, height, numberOfMines)
-            return Cells(cells).apply { NearbyMineCounter.count(this) }
-        }
-    }
 }
